requirements.txt
# Core dependencies
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.20.0
numpy>=1.21.0
opencv-python>=4.5.0
pillow>=8.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
pandas>=1.3.0
tqdm>=4.62.0
PyYAML>=5.4.0
scikit-learn>=1.0.0

# Vision-Language Models
timm>=0.6.0
clip-by-openai>=1.0
lavis>=1.0.0

# Evaluation metrics
pycocotools>=2.0.4
nltk>=3.7
rouge-score>=0.1.2
bert-score>=0.3.12

# Development and testing
pytest>=6.2.0
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0
pre-commit>=2.15.0

# Documentation
sphinx>=4.3.0
sphinx-rtd-theme>=1.0.0

# Utilities
requests>=2.26.0
jsonlines>=2.0.0
h5py>=3.6.0